# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_health",
        "summary": "Health check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "List all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserListItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/{id}/promote-admin": {
      "patch": {
        "operationId": "UsersController_promoteToAdmin",
        "summary": "Promote user to admin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Get current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Logistic API",
    "description": "API для системи управління пасажирськими перевезеннями",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "example": "ok"
          },
          "database": {
            "type": "string",
            "enum": [
              "connected",
              "disconnected"
            ],
            "example": "connected"
          },
          "message": {
            "type": "string",
            "example": "Connection refused"
          }
        },
        "required": [
          "status",
          "database"
        ]
      },
      "UserListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "phone": {
            "type": "string",
            "example": "+380501234567"
          },
          "first_name": {
            "type": "string",
            "example": "Іван"
          },
          "last_name": {
            "type": "string",
            "example": "Петренко"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "driver"
            ],
            "example": "driver"
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "phone",
          "first_name",
          "last_name",
          "role",
          "created_at"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+380501234567"
          },
          "first_name": {
            "type": "string",
            "example": "Іван"
          },
          "last_name": {
            "type": "string",
            "example": "Петренко"
          }
        }
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "phone": {
            "type": "string",
            "example": "+380501234567"
          },
          "first_name": {
            "type": "string",
            "example": "Іван"
          },
          "last_name": {
            "type": "string",
            "example": "Петренко"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "driver"
            ],
            "example": "driver"
          }
        },
        "required": [
          "id",
          "phone",
          "first_name",
          "last_name",
          "role"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+380501234567"
          },
          "first_name": {
            "type": "string",
            "example": "Іван"
          },
          "last_name": {
            "type": "string",
            "example": "Петренко"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "phone",
          "first_name",
          "last_name",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          }
        },
        "required": [
          "access_token",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+380501234567"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      }
    }
  }
}